requirements.txt
# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.0
pytest-rerunfailures>=12.0

# Stagehand browser automation
stagehand>=0.5.0

# Environment management
python-dotenv>=1.0.0

# Data validation (required by stagehand)
pydantic>=2.0.0

